requirements.txt
flask~=2.3.3
psycopg2~=2.9.7
Flask-PostgreSQL~=1.1.0
Flask-SQLAlchemy~=3.0.5
Flask-WTF~=1.1.1
WTForms~=3.0.1
email-validator~=2.0.0


